Flow apply coupons in cart of customer

+++ src/main/java/com/module5/team2/repository/OrderItemRepository.java
@@ -3,5 +3,6 @@ package com.module5.team2.repository;
 import com.module5.team2.entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+//    boolean existsByCouponCode(String couponCode);
 }
